# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Call Assistant API
  description: |
    AI-powered call assistant for NuVance Labs that handles incoming calls, analyzes caller intent using Claude AI, and routes calls appropriately.
    
    This API includes both Twilio webhook endpoints (called by Twilio) and dashboard endpoints (called by the admin frontend).
  version: 1.0.0
  contact:
    name: NuVance Labs
    email: michael@nuvancelabs.com
servers:
  - url: https://your-app.vercel.app
    description: Production server
  - url: http://localhost:3000
    description: Development server
paths:
...
  /api/04-auth/login:
    post:
      summary: Admin authentication
      description: |
        Authenticates admin user with hardcoded credentials.
        Creates session cookie on successful login.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "admin"
                password:
                  type: string
                  example: "your_secure_password"
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Login successful"
          headers:
            Set-Cookie:
              description: Authentication session cookie
              schema:
                type: string
                example: "auth-session=uuid; HttpOnly; Path=/; Expires=..."
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid credentials"
...
  /api/05-dashboard/calls:
    get:
      summary: Fetch call history
      description: |
        Returns paginated list of recent calls with details.
        Requires admin authentication.
      tags:
        - Dashboard
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
        - name: limit
          in: query
          description: Number of calls per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          example: 20
      responses:
        '200':
          description: Call history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CallLog'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 20
                      hasMore:
                        type: boolean
                        example: true
        '401':
          description: Unauthorized - invalid or missing session
        '500':
          description: Server error
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth-session
  schemas:
    CallLog:
      type: object
      properties:
        id:
          type: integer
          description: Unique call record ID
          example: 123
        callSid:
          type: string
          description: Twilio call identifier
          example: "CA1234567890abcdef1234567890abcdef"
        from:
          type: string
          description: Caller's phone number
          example: "+1234567890"
        to:
          type: string
          description: Called phone number
          example: "+1987654321"
        callerName:
          type: string
          nullable: true
          description: Caller's name if captured
          example: "John Smith"
        intent:
          type: string
          nullable: true
          description: AI-classified intent
          enum: [emma_request, sales_general, business_general, nonsense, unclear]
          example: "sales_general"
        transcript:
          type: string
          nullable: true
          description: Call transcript
          example: "I would like to speak to Emma about pricing"
        outcome:
          type: string
          description: Call outcome
          enum: [connected, routing_failed, filtered, message_taken, completed]
          example: "connected"
        duration:
          type: integer
          nullable: true
          description: Call duration in seconds
          example: 120
        status:
          type: string
          description: Final call status
          example: "completed"
        timestamp:
          type: string
          format: date-time
          description: Call timestamp
          example: "2025-09-04T15:30:00Z"
        recordingUrl:
          type: string
          nullable: true
          description: Recording URL if available
          example: "https://api.twilio.com/2010-04-01/Accounts/.../Recordings/..."
...